Comienzo de la integracion
Se agregaron las 5 etapas claramente diferenciadas.
La intención posterior es realizar un merge entre lo modificado en las ramas 1, 2, 3 y 5.
