+++ Proyecto/Etapa3.py
@@ -1,0 +1,5 @@
+"""
+Etapa 3: Rosco
+El responsable de esta etapa es Santiago Testa.
+El responsable de su revision es Tomas Galluccio.
+"""
