+++ Proyecto/Etapa5.py
@@ -1,0 +1,5 @@
+"""
+Etapa 5: Puntaje
+El responsable de esta etapa es Valentino Ceniceros.
+El responsable de su revision es Emilio Ontiveros.
+"""
